requirements.txt
streamlit==1.31.0
groq==0.4.2
google-generativeai==0.3.2
python-dotenv==1.0.0
rich==13.7.0